refactor: move related server actions from projects to projects/members folder.
