+++ src/app/projects/[projectId]/_components/add-member/AddMemberInput.tsx
@@ -12,7 +12,7 @@ import { z } from 'zod';
 import ErrorMessage from '@/components/shared/ErrorMessage';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
-import { addMemberToProject } from '@/lib/actions/project/add-member-to-project';
+import { addMemberToProject } from '@/lib/actions/project/members/add-member-to-project';
 import { useDebounce } from '@/lib/hooks/useDebounce';
 import { useGetUsersByEmailQuery } from '@/lib/store/api';
 import { TUser } from '@/lib/types/tables/user';
